backend/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.27.1
SQLAlchemy==2.0.27
pydantic==2.6.1
pydantic-settings==2.1.0
python-multipart==0.0.9
reportlab==4.0.9
python-jose[cryptography]==3.3.0
passlib==1.7.4
bcrypt==4.2.0
playwright==1.48.0
Jinja2==3.1.4
email-validator==2.1.0
pytest==8.0.0
pytest-asyncio==0.23.4
httpx==0.27.0
pytest-cov==4.0.0
factory-boy==3.3.0
alembic==1.13.1
slowapi==0.1.9
redis==5.0.3
websockets==12.0
python-socketio==5.11.0
